backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
groq>=0.4.1
pydantic>=2.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
spacy>=3.7.2
networkx>=3.2
plotly>=5.17.0
folium>=0.14.0
reportlab>=4.0.4
python-docx>=1.1.0
sqlalchemy>=2.0.23
asyncpg>=0.29.0
redis>=5.0.1
